Carpeta Cliente y Servidor
Cree la carpeta cliente y servidor y puse todo lo de la iteracion 1 en cliente.
La carpeta Servidor está vacia
