+++ Cliente/bin/modelo/Contacto/IFuncionalidadContacto.class
