+++ Cliente/bin/modelo/usuario/IConectar.class
