+++ Cliente/bin/modelo/usuario/IFuncionalidadUsuario.class
