+++ Cliente/bin/vista/INotificable.class
