+++ Cliente/src/main/Main.java
@@ -1,0 +1,12 @@
+package main;
+
+import controlador.Controlador;
+
+public class Main {
+
+	public static void main(String[] args) {
+		Controlador controlador = new Controlador();
+		controlador.Iniciar();
+	}
+
+}
