+++ Cliente/src/modelo/Contacto/IFuncionalidadContacto.java
