+++ Cliente/src/modelo/Contacto/IVerConversacion.java
