+++ Cliente/src/modelo/usuario/IFuncionalidadUsuario.java
