+++ Cliente/src/module-info.java
@@ -1,0 +1,9 @@
+/**
+ * 
+ */
+/**
+ * 
+ */
+module Itera1 {
+	requires java.desktop;
+}
