+++ Cliente/src/vista/INotificable.java
@@ -1,0 +1,5 @@
+package vista;
+
+public interface INotificable {
+	void OnNuevoMensajeRecibido();
+}
