--- src/main/Main.java
@@ -1,12 +1,0 @@
-package main;
-
-import controlador.Controlador;
-
-public class Main {
-
-	public static void main(String[] args) {
-		Controlador controlador = new Controlador();
-		controlador.Iniciar();
-	}
-
-}
