--- src/module-info.java
@@ -1,9 +1,0 @@
-/**
- * 
- */
-/**
- * 
- */
-module Itera1 {
-	requires java.desktop;
-}
