--- src/vista/INotificable.java
@@ -1,5 +1,0 @@
-package vista;
-
-public interface INotificable {
-	void OnNuevoMensajeRecibido();
-}
